day1 part1

+++ src/main.zig
@@ -9,26 +9,26 @@ pub fn main() !void {
     defer f.close();
 
     var reader = f.reader();
-    var leftList = std.ArrayList(usize).init(allocator);
-    var rightList = std.ArrayList(usize).init(allocator);
+    var leftList = std.ArrayList(isize).init(allocator);
+    var rightList = std.ArrayList(isize).init(allocator);
     defer leftList.deinit();
     defer rightList.deinit();
 
-    while (try reader.readUntilDelimiterOrEofAlloc(allocator, '\n', std.math.maxInt(usize))) |line| {
+    while (try reader.readUntilDelimiterOrEofAlloc(allocator, '\n', std.math.maxInt(isize))) |line| {
         defer allocator.free(line);
         var it = std.mem.splitSequence(u8, line, "   ");
         if (it.next()) |x| {
-            try leftList.append(try std.fmt.parseInt(usize, x, 10));
+            try leftList.append(try std.fmt.parseInt(isize, x, 10));
         }
         if (it.next()) |x| {
-            try rightList.append(try std.fmt.parseInt(usize, x, 10));
+            try rightList.append(try std.fmt.parseInt(isize, x, 10));
         }
     }
 
-    std.mem.sort(usize, leftList.items, {}, comptime std.sort.asc(usize));
-    std.mem.sort(usize, rightList.items, {}, comptime std.sort.asc(usize));
+    std.mem.sort(isize, leftList.items, {}, comptime std.sort.asc(isize));
+    std.mem.sort(isize, rightList.items, {}, comptime std.sort.asc(isize));
 
-    var distance: usize = 0;
+    var distance: u64 = 0;
     for (leftList.items, 0..) |leftValue, index| {
         const rightValue = rightList.items[index];
         distance += @abs(leftValue - rightValue);
